# Improve listStatus 5X performance in some scenarios