--- a/core/client/hdfs/src/main/java/alluxio/hadoop/AbstractFileSystem.java
+++ b/core/client/hdfs/src/main/java/alluxio/hadoop/AbstractFileSystem.java
@@ -598,7 +598,8 @@ public abstract class AbstractFileSystem extends org.apache.hadoop.fs.FileSystem
     List<URIStatus> statuses;
     try {
       ListStatusPOptions listStatusPOptions = ListStatusPOptions.getDefaultInstance().toBuilder()
-          .setNoNeedUseMountInfo(true).build();
+          .setNoNeedUseMountInfo(
+              mAlluxioConf.getBoolean(PropertyKey.USER_HDFS_LISTSTATUS_MOUNTINFO_DISABLE)).build();
       statuses = mFileSystem.listStatus(uri, listStatusPOptions);
     } catch (FileDoesNotExistException e) {
       throw new FileNotFoundException(getAlluxioPath(path).toString());