--- a/core/common/src/main/java/alluxio/conf/PropertyKey.java
+++ b/core/common/src/main/java/alluxio/conf/PropertyKey.java
@@ -6050,6 +6050,14 @@ public final class PropertyKey implements Comparable<PropertyKey> {
               + "when Alluxio workers are required but not ready.")
           .setScope(Scope.CLIENT)
           .build();
+  public static final PropertyKey USER_HDFS_LISTSTATUS_MOUNTINFO_DISABLE =
+      booleanBuilder(Name.USER_HDFS_LISTSTATUS_MOUNTINFO_DISABLE)
+          .setDefaultValue(true)
+          .setIsHidden(true)
+          .setDescription("When a hdfs client listStatus a file, Whether not to use mount info")
+          .setConsistencyCheckLevel(ConsistencyCheckLevel.IGNORE)
+          .setScope(Scope.CLIENT)
+          .build();
   public static final PropertyKey USER_LOCAL_READER_CHUNK_SIZE_BYTES =
       dataSizeBuilder(Name.USER_LOCAL_READER_CHUNK_SIZE_BYTES)
           .setDefaultValue("8MB")
@@ -8591,6 +8599,8 @@ public final class PropertyKey implements Comparable<PropertyKey> {
     public static final String USER_FILE_WRITE_INIT_MAX_DURATION =
         "alluxio.user.file.write.init.max.duration";
     public static final String USER_HOSTNAME = "alluxio.user.hostname";
+    public static final String USER_HDFS_LISTSTATUS_MOUNTINFO_DISABLE =
+        "user.hdfs.liststatus.mountinfo.disable";
     public static final String USER_LOCAL_READER_CHUNK_SIZE_BYTES =
         "alluxio.user.local.reader.chunk.size.bytes";
     public static final String USER_LOCAL_WRITER_CHUNK_SIZE_BYTES =