### What changes are proposed in this pull request?
 Improve listStatus 5X performance in some scenarios.

### Why are the changes needed?
Please clarify why the changes are needed. For instance,
Under the scenario:
1. Use Hadoop compatible system to access Alluxio(listStatus)
2. There are many mount point such as more than 500.
3. There are more than 2000 files in a directory.

The PathUtils.hasPrefix (comes from MountTable.getMountPoint)method will be called at least 10w (500 * 2000) times.
But actually we don't need the information of mount point.
The test can be reduced from about 400ms to 70ms under the master branch.
The test can be reduced from about 700ms to about 100ms under 2.7.1 branch.

### Does this PR introduce any user facing changes?

Please list the user-facing changes introduced by your change, including
  no