@flaming-archer 2023-02-16
> Thanks to @flaming-archer for submitting PR and contributing to Alluxio.

Yes ,I think this is a very useful feature. For example, to speed up presto, we do not need use the information about mount points. Even if there are not hundreds of mount points ,for example, 10，2000 subdirectories can reduce 10*2000 =2w calls.