@huanghua78 2023-02-16
> I think "Java 8 Fault Tolerant Integration Tests / modules: (alluxio.server.ft.journal.raft.**,alluxio.server.ft.journal.ufs.**) (pull_request) Failing after 11m" and "Java 8 Integration Tests / modules: (alluxio.client.fs.concurrent.**,alluxio.client.fs.io.**) (pull_req" are not comes from my pr. Both are failed because of dependecny not found, but I do not change any pom.xml

Some of these failures are not from this PR. They are some sporadic issues in the tests themselves. 
Let me schedule to re-run the failed tests.