@flaming-archer 2023-02-23
@jiacheliu3 Hi, I have changed my code , pls take a look at it. 

And another question can we set `USER_HDFS_CLIENT_EXCLUDE_MOUNT_INFO_ON_LIST_STATUS` default  **true**.