@jiacheliu3 2023-02-26
> @jiacheliu3 Hi, I have changed my code , pls take a look at it.
> 
> And another question can we set `USER_HDFS_CLIENT_EXCLUDE_MOUNT_INFO_ON_LIST_STATUS` default **true**.

I don't think it's safe to set `USER_HDFS_CLIENT_EXCLUDE_MOUNT_INFO_ON_LIST_STATUS=true` because the `ufsPath` field is commonly used. I believe the alluxio client will `listStatus` then pass the `FileInfo` to the alluxio worker, so the worker will know what UFS file to read if the cache block does not exist. 

Right now I think the best way to serve all users is to optimize mount resolution code.